+++ CHANGELOG.md
@@ -7,6 +7,8 @@ ChangeLog
 
 - REFACTOR: refactor "t2t-openai" and "t2t-ollama" into "t2t-spellcheck" and "t2t-translate"
 - REFACTOR: rename "t2t-transformers" to "t2t-opus" and reduce its functionality to OPUS only
+- FEATURE: added "a2t-google" node for Google Cloud Speech-to-Text processing
+- FEATURE: added "t2a-google" node for Google Cloud Text-to-Speech processing
 - FEATURE: added "t2t-punctuation" text-to-text punctuation node
 - FEATURE: added "t2t-summary" text-to-text summarization node
 - FEATURE: added "t2t-profanity" text-to-text profanity filtering node